requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
firebase-admin>=6.6.0
pydantic>=2.12.0
python-dotenv>=1.0.1
httpx>=0.28.1
